+++ src/background/main.rs
@@ -25,10 +25,13 @@ use std::io::{BufWriter, Write};
 use color_eyre::owo_colors::OwoColorize;
 use error_handler::Result;
 use exposed::register_invoke_handler;
-use modules::{cli::{
-    application::{handle_cli_info, SEELEN_COMMAND_LINE},
-    infrastructure::Client,
-}, tray::application::force_tray_overflow_creation};
+use modules::{
+    cli::{
+        application::{handle_cli_info, SEELEN_COMMAND_LINE},
+        infrastructure::Client,
+    },
+    tray::application::force_tray_overflow_creation,
+};
 use plugins::register_plugins;
 use seelen::SEELEN;